sincID maxID int64

+++ comments_by_me_test.go
@@ -27,8 +27,8 @@ func TestCommentsByMe(t *testing.T) {
 	if err != nil {
 		t.Fatal(err)
 	}
-	sinceID := 0
-	maxID := 0
+	sinceID := int64(0)
+	maxID := int64(0)
 	count := 50
 	page := 1
 	filterBySource := 0
